requirements.txt

numpy
faiss-cpu
requests
fastapi
pydantic
azure-core
azure-ai-formrecognizer
azure-ai-documentintelligence